EllipticCurves/DUAL_EC_DRBG_CHALL/requirements.txt

cryptography==43.0.1
fastecdsa
pwntools==4.13.0
pycryptodome==3.20.0

